+++ logs/reqLog.log
@@ -6093,3 +6093,4 @@
 20230117	16:39:54	abd54756-9f47-468f-bac8-74764c2244a2	GET	/users	http://127.0.0.1:5173
 20230117	18:33:47	53cba9ac-3035-4769-a127-a84e5b930a8d	GET	/auth/refresh	http://127.0.0.1:5173
 20230117	18:33:47	fc3fc5c4-4322-4b23-863c-e16a43dcd088	GET	/auth/refresh	http://127.0.0.1:5173
+20230119	09:17:22	8c1adda8-814d-4b78-a44e-833163148571	GET	/act	undefined
