+++ routes/actRoute.js
@@ -1,14 +1,14 @@
 const express = require('express')
 const router = express.Router()
 const actController = require('../controllers/actController')
-const verifyJWT = require('../middleware/verifyJWT')
+//const verifyJWT = require('../middleware/verifyJWT')
 
-router.use(verifyJWT)
+//router.use(verifyJWT)
 
 router.route('/')
-    .get(actController.getAllAct)
+    .get(actController.getAllActs)
     .post(actController.createNewAct)
-    .patch(actController.updateNoteAct)
-    .delete(actController.deleteNoteAct)
+    .patch(actController.updateAct)
+    .delete(actController.deleteAct)
 
 module.exports = router
